fix: move packed broker-worker check into its own verifier script
The runtime-bootstrap verifier has a unit test with a bin-only fixture, so
the embedded-wasm check broke it in CI. The check now lives in
verify-packed-broker-worker.mjs, run as a second step against the same
extracted tarball in the publish workflow.
